requirements.txt
auto-py-to-exe
customtkinter==5.2.0
dependency-injector
psycopg2-binary
python-dotenv